LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigation-methods/traverseTo-after-adding-iframe-expected.txt



PASS navigation.traverseTo() should work in an iframe that is not present in all history entries
